update price in phone model

+++ src/main/java/com/app/gadgetblitz/controller/PhoneController.java
@@ -4,14 +4,10 @@ import com.app.gadgetblitz.dto.PhoneFullDto;
 import com.app.gadgetblitz.dto.PhoneSimpleDto;
 import com.app.gadgetblitz.service.PhoneService;
 import lombok.AllArgsConstructor;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
 import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
@@ -22,14 +18,6 @@ public class PhoneController {
 
     MongoTemplate mongoTemplate;
     PhoneService phoneService;
-
-//    @GetMapping()
-//    public ResponseEntity<List<PhoneSimpleDto>> getAllPhones() {
-//        List<PhoneSimpleDto> phones = phoneService.findAll();
-//        if (phones.isEmpty())
-//            return ResponseEntity.notFound().build();
-//        return ResponseEntity.ok(phones);
-//    }
 
     @GetMapping()
     public ResponseEntity<List<PhoneSimpleDto>> getAllPhones(
@@ -43,8 +31,6 @@ public class PhoneController {
 
         return ResponseEntity.ok(phonesPage);
     }
-
-
 
     @GetMapping("/{id}")
     public ResponseEntity<Optional<PhoneFullDto>> getPhoneById(@PathVariable String id) {
@@ -62,8 +48,8 @@ public class PhoneController {
             @RequestParam(required = false) Double sizeMax,
             @RequestParam(required = false) Integer storageMin,
             @RequestParam(required = false) Integer storageMax,
-            @RequestParam(required = false) BigDecimal priceMin,
-            @RequestParam(required = false) BigDecimal priceMax,
+            @RequestParam(required = false) Double priceMin,
+            @RequestParam(required = false) Double priceMax,
             @RequestParam(required = false) Integer cameraBackMin,
             @RequestParam(required = false) Integer cameraBackMax) {
